# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: 'MSP-MIU Website API'
  description: "REST API for the MSP Tech Club MIU platform (members, events, CMS, competitions, quizzes).\n\nCall `POST /auth/login` (or `/users/login`) — the JWT is applied to **Authorize** automatically.\n"
  version: 1.0.0
  contact:
    name: 'MSP Tech Club MIU'
    url: 'https://msp-miu.tech'
servers:
  -
    url: 'https://msp-miu.tech/api'
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SeasonId:
      name: season_id
      in: query
      schema:
        oneOf:
          -
            type: integer
          -
            type: string
            enum:
              - current
              - all
      description: 'Season filter — omit/current, numeric id, or all'
...
security: []
paths:
...
  '/admin/teams/{teamId}/members/{teamMemberId}':
    put:
      tags:
        - Admin
      summary: 'Update team member'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: teamId
          in: path
          required: true
          schema:
            type: string
        -
          name: teamMemberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
    delete:
      tags:
        - Admin
      summary: 'Remove team member'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: teamId
          in: path
          required: true
          schema:
            type: string
        -
          name: teamMemberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed
...
  /members:
    get:
      tags:
        - Members
      summary: 'List members'
      security:
        -
          bearerAuth: []
      parameters:
        -
          $ref: '#/components/parameters/SeasonId'
      responses:
        '200':
          description: List
    post:
      tags:
        - Members
      summary: 'Create member'
      security:
        -
          bearerAuth: []
      responses:
        '201':
          description: Created
  '/members/{id}':
    get:
      tags:
        - Members
      summary: 'Get member'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member
    put:
      tags:
        - Members
      summary: 'Update member'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
    delete:
      tags:
        - Members
      summary: 'Delete member'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...